Fixed errors

+++ src/test/java/com/itesm/application/usecase/UpdateUserProfileUseCaseTest.java
@@ -68,7 +68,7 @@ public class UpdateUserProfileUseCaseTest {
         assertEquals("NewName", result.getName());
         assertEquals("NewLastName1", result.getLastName1());
         assertEquals("NewLastName2", result.getLastName2());
-        assertEquals(Integer.valueOf(35), result.getAge());
+        assertEquals((byte) 35, result.getAge());
         assertNotNull(result.getSuburb());
 
         verify(userRepository).findDomainById(1L);
@@ -106,7 +106,7 @@ public class UpdateUserProfileUseCaseTest {
         assertNotNull(result);
         assertEquals("OnlyName", result.getName());
         assertEquals("Existing2", result.getLastName2());
-        assertEquals(Integer.valueOf(30), result.getAge());
+        assertEquals((byte) 30, result.getAge());
         assertNull(result.getSuburb());
 
         ArgumentCaptor<User> captor = ArgumentCaptor.forClass(User.class);
@@ -144,7 +144,7 @@ public class UpdateUserProfileUseCaseTest {
         assertNotNull(result);
         assertEquals("Citizen", result.getName());
         assertEquals("Existing2", result.getLastName2());
-        assertEquals(Integer.valueOf(30), result.getAge());
+        assertEquals((byte) 30, result.getAge());
         assertNull(result.getSuburb());
 
         ArgumentCaptor<User> captor = ArgumentCaptor.forClass(User.class);
